Update osi_get_bytes_used to support platforms without mallinfo()
mallinfo() is the only glibc-specific function used in all of
swish. This change allows platforms without mallinfo() or another good
way to measure memory usage to just return 0. Previous swish projects
have not relied on (osi_get_bytes_used) for anything except logging
and leak detection so this should be low-impact.

An alternative approach would be to read /proc/self/statm but it would
be much slower, less accurate, and could be implemented in scheme if
required.