swish-build: support --rtlib for --library
The --rtlib option can now be used with the --library option to prevent
whole-program optimization from duplicating code that will be provided
separately at run time.

We use library-search-handler to control whole-program optimization,
which means we require Chez Scheme version 9.5.2 or later.

We now visit the libraries specified by --rtlib before compiling
the input file; not doing this was a bug that could have let us
fail to find a library present only as an --rtlib binary or let
us incorporate a library for which we had a source or .wpo file
available when we should have left it as a run-time library.