configure: fix --scheme=<path-to-scheme>
Common uses for the ./configure --scheme=<path-to-scheme> option are to
build against the version in a particular Chez Scheme workarea or to
build against a version installed by Chez Scheme's Mf-install under a
particular prefix. If we have trouble running the specified scheme,
we try setting SCHEMEHEAPDIRS to cover these directories.