swish-start: import available swish libraries
For convenience, we make swish library exports visible at the repl and in
scripts, much as Chez Scheme does for its libraries. We used to do this
via (import (swish imports)), but that breaks stand-alone applications
that build in a subset of the swish libraries. Instead, we selectively
import libraries that are present and appear to be swish libraries.