swish-build: generate and retain expected object file for --library
During the compile-library phase, we now use the expected object filename
if it can be determined by consulting library-extensions.  Otherwise, we
fall back on a temporary filename as before. After compile-whole-library,
we remove the object file only when using a temporary filename.

This means that, by default, swish-build --library generates a ".so" file
before generating the -o file via whole-program optimization. This is useful
since import looks for the ".so" file (unless prevented by --rtlib). We used
to hand import the ".wpo" file, but the ".so" file has the advantage of already
been compiled.