testing: fix error during "make coverage"
The broken test incorrectly expected (and ignored) a single return
value from swish-start, when it actually returns the values, if any,
passed to exit.

When a test fails in "make coverage" but not in "make test", the
difference may be due to the fact that we disable cp0 for coverage
runs. That can reveal places where we inadvertently rely on cp0 to
optimize away an unreferenced binding that expects a particular
number of values. That's what was happening in the broken test.