add json:cells, json:object?, json:set!, json:size
We can't derive our own record type for json records since the native
hashtable record types are sealed.  Instead we install and check for a
specific equivalence function as evidence that a hashtable was contructed
by json:make-object.

Where it is safe, we use #3% versions of symbol-hashtable accessors
to bypass symbol-hashtable checks that would otherwise balk at the
unrecognized equivalence function. Note that the inlined Chez Scheme
library functions use eq? and ignore the equivalence function we provide.