Revert "use a more efficient expansion for certain match forms"
This reverts commit df3a2e5f2cdbecb7c40bbdaad19404710a76a240.

The criteria used there to justify the transformation are not strong
enough.  In particular, literals and ,@var patterns are effectively
guards, but are not detected as such. Patterns containing pairs also
introduce more conditional control flow. All of these cases can lead
to duplicating the code of the second match clause.

These are okay:
 (expand '(match input [,@var 'ok] [,other "don't duplicate"]))
 (expand '(match input [#(,x) x] [,other "don't duplicate"]))
 (expand '(match input [#(,x ,y) x] [,other "don't duplicate"]))

These result in duplication:
 (expand '(match input [(,x) x] [,other "don't duplicate"]))
 (expand '(match input [#(,@x) x] [,other "don't duplicate"]))
 (expand '(match input [#(,x y) x] [,other "don't duplicate"]))

We could do more to scrutinize the pattern, but it is not worth
doing at this point.