fix valgrind --leak-check=full errors
Calling exit() gives functions registered via atexit() a chance to run
on normal process termination. In particular, this lets libuv run its
threadpool cleanup code, which makes valgrind happy. However, if some
thread in the pool is blocked on I/O, we may block indefinitely when
the cleanup code tries to join with that thread.

If the libuv event loop doesn't shut down cleanly, say because I/O
is pending, we call _exit() instead to shut down forcibly.

Of course, we also need to unhook the signal handlers and the timer
and be sure the close gets a chance to run.