calling osi_send_request from the main thread returns UV_EPERM
Return UV_EPERM if osi_send_request is called on the main thread
since it is not safe to call osi_add_callbacks unless we are in
the dynamic context of the event-loop's call to osi_get_callbacks.

Update the documentation for osi_add_callback_list (etc.) to
clarify the condition under which they may safely be called.
Calling on the main thread is necessary, but not sufficient.