use absolute path in repl.ss to cope with csv9.5.3 delayed revisit
After csv9.5.3 commit 88501743001393fa82e89c90da9185fc0086fbcb
the run-mat script started failing with a confusing error message
saying it couldn't find one of mat.so, profile.so, or testing.so,
depending on which of them were already compiled.

This is because we import the libraries, which now triggers only
a 'visit' if already compiled, before changing the working directory
and calling run-mats, which now triggers a 'revisit' that fails if
the library was originally imported via a relative path.