make stack useful for (throw 'timeout)
Add call to void within test thunks so we get useful stack for mats
that do (receive (after 1000 (throw 'timeout)) ...).

Use raise instead of throw within $isolate-mat so we don't add a stack
from the test harness, but (raise e) in place of (raise reason) so we
propagate stack from exception in the code being tested.