retain more debugging context
In general, use throw except where we trap an exception while
calling user code. There it may be better to match `(catch ,r ,e)
and (raise e) so we propagate the detailed exit reason, if available,
without attributing reason r to our code as would happen if we did
(throw r) or adding a stack that is likely noise if we did (throw e).

We use limit-stack in a couple places where we call user code to help
focus debugging attention on the user code.