fix "unexpected return from reset handler"
If a script called app:start and then caused an exception, we could
get a flood of "unexpected return from reset handler" when the initial
reset-handler called application:shutdown, which found and killed the
application object, but failed to wait for the application to terminate.
Note that new-cafe installs its own reset-handler, so interactive sessions
were typically unaffected.