add repl-level parameter to indicate when REPL is available
We can't rely on Chez Scheme's #%$cafe since it doesn't understand
Swish processes. If a process that is not running a new-cafe calls
app-exception-handler we don't want to offer "Type (debug) to enter
the debugger." since there may be no way to do so. In particular, if
an error in pid1 shuts down the application, we don't want to offer the
missed chance to debug on the way out.