fix: shut down on error in process linked to initial process
If an error occurred in a process linked to the initial process, the
system did not shut down since @kill takes out the process out before
it can run the base-exception-handler or reset-handler.

To fix this, we add a process to montior the initial process and shut down
the application on failure. @kill explicitly checks for some essential
system proceses, but no one was watching out for the initial process.