fix: an error in osi_get_ip_address inside a callback crashes the event-loop
The direct call to osi_get_ip_address inside the callback procedures
for connect-tcp and listen-tcp can fail. When it does, the event-loop
crashes the entire Swish process.

Failing to get the ip address for the connection does not invalidate
the connection. We try to use a reasonable name if available.