add make-fault/no-cc
The make-fault/no-cc procedure behaves like make-fault, except that it
omits the current continuation.

When a gen-server's handle-call reports a fault to the gen-server:call
site, it can use make-fault/no-cc to omit the continuation within the
gen-server code and draw attention instead to the caller that supplied
invalid arguments.