breaking change: add and log global-process-id
Use the same process notation in pcb record-writer and log-db coerce.

Previously it was possible for console-event-handler messages and
database error details to contain references to a "#<process 7>" that
had no relation to the various database tables that might contain 7 in
a pid column.

Logging a global-process-id means that database pids now correspond
predictably to run-time pids. In particular, database pids no longer
depend on the order of evaluation for calls to coerce. For example,
it had been possible for a supervisor to be assigned a higher pid in
the database than one of its child processes.

Breaking:
- database columns containing process ids are now text, not integer
- the printed representation of processes is changed