add (response-timeout) to http:options
`http:respond` and `http:respond-file` use the default 5 second gen-server
timeout which is not enough for slow networks or large amounts of data.

Adding an optional timeout argument to each of these functions helps,
but does not cover the static file handler buried in the http-cache server.

Adding (response-timeout) allows Swish applications to override this
behavior when configuring the server.

The testing uses a buffer size of 2^29-1 bytes to support 32-bit
systems and a 15 ms timeout. Current 64-bit systems can transfer this
near 50 ms, so we have some room for timing error.