fix: by-pass filters in inherit-parameters
When inherit-parameters installs parameter values in the spawned
process, we now avoid calling the filter, if any, provided in the call
to make-inherited-parameter. Any filtering has already been applied when
these parameters are set and we want to avoid the additional overhead in
spawn and spawn&link. This is particularly important for the inherited
app:path parameter, where the filter makes an expensive call to
get-real-path.