extend sqlite:execute
sqlite:execute now calls the new osi_bind_statement_bindings function
directly. If the bindings argument is not a marshaled bindings
record instance, it first calls sqlite:marshal-bindings and calls
sqlite:unmarshal-bindings on exit. As a result, sqlite:execute now accepts
as bindings: a list, a vector, or a marshaled bindings record instance.

Co-authored-by: Chris Payne <capayne@beckman.com>