add osi_bulk_execute and sqlite:bulk-execute
In Swish versions up through 2.0.2, a flood of db:log calls can cause
the db gen-server process to fall behind and slow the application. This
deluge can create an out-of-memory condition that crashes the
application.

To remedy this, we add osi_bulk_execute, which takes a vector of SQLite
statements and a vector of bindings, and requests that the database
worker thread do all of the SQLite binding and stepping of statements
before returning. The db gen-server now flushes its queue of db:log
statements using osi_bulk_execute.

We define a "marshaled bindings" C struct and a corresponding Scheme
record type. Each marshaled bindings record represents a complete
set of values that the database worker thread will bind to the
statement.

In practice, the new code took a specific Swish application from an
out-of-memory failure condition, where the database could not keep
up, to maintaining log performance with less than 50 rows per
transaction, where less than 5 is most frequent.

Co-authored-by: Oscar Waddell <owaddell@beckman.com>