revise define-options
The original implementation was elegant, but it relied on parameters
and had to be used with care.

The new implementation simply builds native records. It supports many
features of define-tuple, such as copy, copy*, and field accessors, yet
adds support for optional fields, default values, predicates, and
filters. The aim is to allow us to extend configurable interfaces
without treating each extension as a breaking change.