allow compiling with cp0 disabled
With cp0 enabled, the compiler inlines the q record accessors and
mutators and omits trap checks in insert-q and remove-q. But with
cp0 disabled, we need to defend $receive against the possibility
of a context switch due to a trap check in remove-q.

Wrapping the remove-q call in no-interrupts may introduce a little
run time overhead, but it is easier to reason about.