lift some work out of pregexp-match-positions-aux
The pregexp-match-positions-aux procedure was burrowing through the entire
regexp to construct an alist of backrefs. That can be inefficient since
pregrexp-match-positions calls it in a loop. Instead, we construct the list
of backref keys at expand time, if possible.

Legacy code that directly manipulated list-structured "compiled regexps"
in the original format can call the pregexp procedure on that data to obtain
a "compiled regexp" in the new form.

Note: be sure to set (print-graph #t) if you expand uses of the (re ...) macro,
write the expanded output to a port, and attempt to read and evaluate the
expanded code. (This practice is neither common nor recommended.)